Test Cases restructuring
