+++ test/UnitTests/PlatformServices.Desktop.Unit.Tests/PlatformServices.Desktop.Unit.Tests.csproj
@@ -64,6 +64,12 @@
     <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\Extensions\ns10ExceptionExtensionsTests.cs">
       <Link>Extensions\ns10ExceptionExtensionsTests.cs</Link>
     </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\Services\ns10MSTestAdapterSettingsTests.cs">
+      <Link>Services\ns10MSTestAdapterSettingsTests.cs</Link>
+    </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\Utilities\ns10DeploymentUtilityTests.cs">
+      <Link>Utilities\ns10DeploymentUtilityTests.cs</Link>
+    </Compile>
     <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.0\Utilities\ns10FileUtilityTests.cs">
       <Link>Utilities\ns10FileUtilityTests.cs</Link>
     </Compile>
@@ -73,23 +79,29 @@
     <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.3\ns13TraceListenerTests.cs">
       <Link>Services\ns13TraceListenerTests.cs</Link>
     </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.3\Services\ns13MSTestSettingsProviderTests.cs">
+      <Link>Services\ns13MSTestSettingsProviderTests.cs</Link>
+    </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.3\Utilities\ns13DeploymentItemUtilityTests.cs">
+      <Link>Utilities\ns13DeploymentItemUtilityTests.cs</Link>
+    </Compile>
+    <Compile Include="..\PlatformServices.Shared.Unit.Tests\netstandard1.3\Utilities\ns13ReflectionUtilityTests.cs">
+      <Link>Utilities\ns13ReflectionUtilityTests.cs</Link>
+    </Compile>
     <Compile Include="AssemblyResolverTests.cs" />
     <Compile Include="Deployment\AssemblyLoadWorkerTests.cs" />
-    <Compile Include="Services\MSTestAdapterSettingsTests.cs" />
     <Compile Include="Services\DesktopReflectionOperationsTests.cs" />
     <Compile Include="Services\DesktopThreadOperationsTests.cs" />
-    <Compile Include="Services\DesktopSettingsProviderTests.cs" />
     <Compile Include="Services\DesktopTestDataSourceTests.cs" />
     <Compile Include="Services\DesktopTestContextImplTests.cs" />
     <Compile Include="Services\DesktopTestDeploymentTests.cs" />
     <Compile Include="Services\DesktopTestSourceHostTests.cs" />
     <Compile Include="Services\DesktopTestSourceTests.cs" />
     <Compile Include="Utilities\AppDomainUtilitiesTests.cs" />
-    <Compile Include="Utilities\DeploymentItemUtilityTests.cs" />
     <Compile Include="Services\DesktopFileOperationsTests.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="Utilities\DeploymentUtilityTests.cs" />
-    <Compile Include="Utilities\ReflectionUtilityTests.cs" />
+    <Compile Include="Utilities\DesktopDeploymentUtilityTests.cs" />
+    <Compile Include="Utilities\DesktopReflectionUtilityTests.cs" />
     <Compile Include="Utilities\VSInstallationUtilitiesTests.cs" />
     <Compile Include="Utilities\XmlUtilitiesTests.cs" />
   </ItemGroup>
@@ -117,7 +129,7 @@
     <ProjectReference Include="..\..\..\src\TestFramework\Extension.Desktop\Extension.Desktop.csproj">
       <Project>{a7ea583b-a2b0-47da-a058-458f247c7575}</Project>
       <Name>Extension.Desktop</Name>
-      <Aliases>FrameworkV2DesktopExtension</Aliases>
+      <Aliases>FrameworkV2Extension</Aliases>
     </ProjectReference>
     <ProjectReference Include="..\..\..\src\TestFramework\MSTest.Core\MSTest.Core.csproj">
       <Project>{7252d9e3-267d-442c-96bc-c73aef3241d6}</Project>
