+++ test/UnitTests/PlatformServices.Desktop.Unit.Tests/Utilities/DesktopDeploymentUtilityTests.cs
@@ -1,8 +1,11 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace MSTestAdapter.PlatformServices.Tests.Utilities
+namespace MSTestAdapter.PlatformServices.Desktop.UnitTests.Utilities
 {
+    extern alias FrameworkV1;
+    extern alias FrameworkV2;
+
     using System;
     using System.Collections.Generic;
     using System.IO;
@@ -14,10 +17,16 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Adapter;
     using Microsoft.VisualStudio.TestPlatform.ObjectModel.Logging;
     using Moq;
-    using Microsoft.VisualStudio.TestTools.UnitTesting;
+    using MSTestAdapter.PlatformServices.Tests.Utilities;
+    using Assert = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
+    using TestClass = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute;
+    using TestInitialize = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestInitializeAttribute;
+    using TestMethod = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
 
     [TestClass]
+#pragma warning disable SA1649 // File name must match first type name
     public class DeploymentUtilityTests
+#pragma warning restore SA1649 // File name must match first type name
     {
         private const string TestRunDirectory = "C:\\temp\\testRunDirectory";
         private const string RootDeploymentDirectory = "C:\\temp\\testRunDirectory\\Deploy";
@@ -53,63 +62,6 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
             this.mocktestExecutionRecorder = new Mock<ITestExecutionRecorder>();
         }
 
-        #region CreateDeploymentDirectories tests
-
-        [TestMethod]
-        public void CreateDeploymentDirectoriesShouldCreateDeploymentDirectoryFromRunContext()
-        {
-            // Setup mocks
-            this.mockRunContext.Setup(rc => rc.TestRunDirectory).Returns(TestRunDirectory);
-            this.mockFileUtility.Setup(fu => fu.GetNextIterationDirectoryName(TestRunDirectory, It.IsAny<string>()))
-                .Returns(RootDeploymentDirectory);
-
-            // Act.
-            this.deploymentUtility.CreateDeploymentDirectories(this.mockRunContext.Object);
-
-            // Assert.
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(RootDeploymentDirectory), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentOutDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix), Environment.MachineName)), Times.Once);
-        }
-
-        [TestMethod]
-        public void CreateDeploymentDirectoriesShouldCreateDefaultDeploymentDirectoryIfTestRunDirectoryIsNull()
-        {
-            // Setup mocks
-            this.mockRunContext.Setup(rc => rc.TestRunDirectory).Returns((string)null);
-            this.mockFileUtility.Setup(fu => fu.GetNextIterationDirectoryName(It.Is<string>(s => s.Contains(TestRunDirectories.DefaultDeploymentRootDirectory)), It.IsAny<string>()))
-                .Returns(RootDeploymentDirectory);
-
-            // Act.
-            this.deploymentUtility.CreateDeploymentDirectories(this.mockRunContext.Object);
-
-            // Assert.
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(RootDeploymentDirectory), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentOutDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix), Environment.MachineName)), Times.Once);
-        }
-
-        [TestMethod]
-        public void CreateDeploymentDirectoriesShouldCreateDefaultDeploymentDirectoryIfRunContextIsNull()
-        {
-            // Setup mocks
-            this.mockFileUtility.Setup(fu => fu.GetNextIterationDirectoryName(It.Is<string>(s => s.Contains(TestRunDirectories.DefaultDeploymentRootDirectory)), It.IsAny<string>()))
-                .Returns(RootDeploymentDirectory);
-
-            // Act.
-            this.deploymentUtility.CreateDeploymentDirectories(null);
-
-            // Assert.
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(RootDeploymentDirectory), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentOutDirectorySuffix)), Times.Once);
-            this.mockFileUtility.Verify(fu => fu.CreateDirectoryIfNotExists(Path.Combine(Path.Combine(RootDeploymentDirectory, TestRunDirectories.DeploymentInDirectorySuffix), Environment.MachineName)), Times.Once);
-        }
-
-        #endregion
-
         #region Deploy tests
 
         [TestMethod]
@@ -121,6 +73,12 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
 
             this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
             this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
 
             Assert.IsFalse(
                 this.deploymentUtility.Deploy(
@@ -132,10 +90,133 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
         }
 
         [TestMethod]
+        public void DeployShouldDeploySourceAndItsConfigFile()
+        {
+            TestRunDirectories testRunDirectories;
+            var testCase = this.GetTestCaseAndTestRunDirectories(DefaultDeploymentItemPath, DefaultDeploymentItemOutputDirectory, out testRunDirectories);
+
+            // Setup mocks.
+            this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
+            this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
+
+            // Act.
+            Assert.IsTrue(
+                this.deploymentUtility.Deploy(
+                    new List<TestCase> { testCase },
+                    testCase.Source,
+                    this.mockRunContext.Object,
+                    this.mocktestExecutionRecorder.Object,
+                    testRunDirectories));
+
+            // Assert.
+            string warning;
+            var sourceFile = Assembly.GetExecutingAssembly().GetName().Name + ".dll";
+            var configFile = Assembly.GetExecutingAssembly().GetName().Name + ".dll.config";
+            this.mockFileUtility.Verify(
+                fu =>
+                fu.CopyFileOverwrite(
+                    It.Is<string>(s => s.Contains(sourceFile)),
+                    Path.Combine(testRunDirectories.OutDirectory, sourceFile),
+                    out warning),
+                Times.Once);
+            this.mockFileUtility.Verify(
+                fu =>
+                fu.CopyFileOverwrite(
+                    It.Is<string>(s => s.Contains(configFile)),
+                    Path.Combine(testRunDirectories.OutDirectory, configFile),
+                    out warning),
+                Times.Once);
+        }
+
+        [TestMethod]
+        public void DeployShouldDeployDependentFiles()
+        {
+            var dependencyFile = "C:\\temp\\dependency.dll";
+
+            TestRunDirectories testRunDirectories;
+            var testCase = this.GetTestCaseAndTestRunDirectories(DefaultDeploymentItemPath, DefaultDeploymentItemOutputDirectory, out testRunDirectories);
+
+            // Setup mocks.
+            this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
+            this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { dependencyFile });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
+
+            // Act.
+            Assert.IsTrue(
+                this.deploymentUtility.Deploy(
+                    new List<TestCase> { testCase },
+                    testCase.Source,
+                    this.mockRunContext.Object,
+                    this.mocktestExecutionRecorder.Object,
+                    testRunDirectories));
+
+            // Assert.
+            string warning;
+
+            this.mockFileUtility.Verify(
+                fu =>
+                fu.CopyFileOverwrite(
+                    It.Is<string>(s => s.Contains(dependencyFile)),
+                    Path.Combine(testRunDirectories.OutDirectory, Path.GetFileName(dependencyFile)),
+                    out warning),
+                Times.Once);
+        }
+
+        [TestMethod]
+        public void DeployShouldDeploySatelliteAssemblies()
+        {
+            TestRunDirectories testRunDirectories;
+            var testCase = this.GetTestCaseAndTestRunDirectories(DefaultDeploymentItemPath, DefaultDeploymentItemOutputDirectory, out testRunDirectories);
+            var assemblyFullPath = Assembly.GetExecutingAssembly().Location;
+            var satelliteFullPath = Path.Combine(Path.GetDirectoryName(assemblyFullPath), "de", "satellite.dll");
+
+            // Setup mocks.
+            this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
+            this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(assemblyFullPath))
+                .Returns(new List<string> { satelliteFullPath });
+
+            // Act.
+            Assert.IsTrue(
+                this.deploymentUtility.Deploy(
+                    new List<TestCase> { testCase },
+                    testCase.Source,
+                    this.mockRunContext.Object,
+                    this.mocktestExecutionRecorder.Object,
+                    testRunDirectories));
+
+            // Assert.
+            string warning;
+
+            this.mockFileUtility.Verify(
+                fu =>
+                fu.CopyFileOverwrite(
+                    It.Is<string>(s => s.Contains(satelliteFullPath)),
+                    Path.Combine(testRunDirectories.OutDirectory, "de", Path.GetFileName(satelliteFullPath)),
+                    out warning),
+                Times.Once);
+        }
+
+        [TestMethod]
         public void DeployShouldNotDeployIfOutputDirectoryIsInvalid()
         {
             TestRunDirectories testRunDirectories;
-            var assemblyFullPath = Assembly.GetEntryAssembly().Location;
+            var assemblyFullPath = Assembly.GetExecutingAssembly().Location;
             var deploymentItemPath = "C:\\temp\\sample.dll";
             var deploymentItemOutputDirectory = "..\\..\\out";
 
@@ -144,6 +225,12 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
             // Setup mocks.
             this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
             this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
 
             // Act.
             Assert.IsTrue(
@@ -181,7 +268,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
         public void DeployShouldDeployContentsOfADirectoryIfSpecified()
         {
             TestRunDirectories testRunDirectories;
-            var assemblyFullPath = Assembly.GetEntryAssembly().Location;
+            var assemblyFullPath = Assembly.GetExecutingAssembly().Location;
 
             var testCase = this.GetTestCaseAndTestRunDirectories(DefaultDeploymentItemPath, DefaultDeploymentItemOutputDirectory, out testRunDirectories);
             var content1 = Path.Combine(DefaultDeploymentItemPath, "directoryContents.dll");
@@ -190,10 +277,16 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
             // Setup mocks.
             this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !(s.EndsWith(".dll") || s.EndsWith(".config"))))).Returns(true);
             this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
             this.mockFileUtility.Setup(
                 fu => fu.AddFilesFromDirectory(DefaultDeploymentItemPath, It.IsAny<bool>())).Returns(directoryContentFiles);
             this.mockFileUtility.Setup(
                 fu => fu.AddFilesFromDirectory(DefaultDeploymentItemPath, It.IsAny<Func<string, bool>>(), It.IsAny<bool>())).Returns(directoryContentFiles);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
 
             // Act.
             Assert.IsTrue(
@@ -216,7 +309,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
                 Times.Once);
         }
 
-        /* [TestMethod]
+        [TestMethod]
         public void DeployShouldDeployPdbWithSourceIfPdbFileIsPresentInSourceDirectory()
         {
             TestRunDirectories testRunDirectories;
@@ -225,6 +318,12 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
             // Setup mocks.
             this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !s.EndsWith(".dll")))).Returns(true);
             this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
             string warning;
             this.mockFileUtility.Setup(fu => fu.CopyFileOverwrite(It.IsAny<string>(), It.IsAny<string>(), out warning))
                 .Returns(
@@ -244,8 +343,8 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
                     testRunDirectories));
 
             // Assert.
-            var sourceFile = Assembly.GetEntryAssembly().GetName().Name + ".dll";
-            var pdbFile = Assembly.GetEntryAssembly().GetName().Name + ".pdb";
+            var sourceFile = Assembly.GetExecutingAssembly().GetName().Name + ".dll";
+            var pdbFile = Assembly.GetExecutingAssembly().GetName().Name + ".pdb";
             this.mockFileUtility.Verify(
                 fu =>
                 fu.CopyFileOverwrite(
@@ -276,6 +375,12 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
             // Setup mocks.
             this.mockFileUtility.Setup(fu => fu.DoesDirectoryExist(It.Is<string>(s => !s.EndsWith(".dll")))).Returns(true);
             this.mockFileUtility.Setup(fu => fu.DoesFileExist(It.IsAny<string>())).Returns(true);
+            this.mockAssemblyUtility.Setup(
+                au => au.GetFullPathToDependentAssemblies(It.IsAny<string>(), It.IsAny<string>(), out this.warnings))
+                .Returns(new string[] { dependencyFile });
+            this.mockAssemblyUtility.Setup(
+                au => au.GetSatelliteAssemblies(It.IsAny<string>()))
+                .Returns(new List<string> { });
             string warning;
             this.mockFileUtility.Setup(fu => fu.CopyFileOverwrite(It.IsAny<string>(), It.IsAny<string>(), out warning))
                 .Returns(
@@ -310,7 +415,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
                     Path.Combine(testRunDirectories.OutDirectory, Path.GetFileName(pdbFile)),
                     out warning),
                 Times.Never);
-        } */
+        }
 
         #endregion
 
@@ -318,7 +423,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
 
         private TestCase GetTestCaseAndTestRunDirectories(string deploymentItemPath, string defaultDeploymentItemOutputDirectoryout, out TestRunDirectories testRunDirectories)
         {
-            var testCase = new TestCase("A.C.M", new System.Uri("executor://testExecutor"), Assembly.GetEntryAssembly().Location);
+            var testCase = new TestCase("A.C.M", new System.Uri("executor://testExecutor"), Assembly.GetExecutingAssembly().Location);
             var kvpArray = new[]
                     {
                         new KeyValuePair<string, string>(
@@ -326,7 +431,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Utilities
                             defaultDeploymentItemOutputDirectoryout)
                     };
             testCase.SetPropertyValue(DeploymentItemUtilityTests.DeploymentItemsProperty, kvpArray);
-            var currentExecutingFolder = Path.GetDirectoryName(Assembly.GetEntryAssembly().Location);
+            var currentExecutingFolder = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
 
             testRunDirectories = new TestRunDirectories(currentExecutingFolder);
