+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.0/ns10FileOperationsTests.cs
@@ -18,7 +18,7 @@ namespace MSTestAdapter.PlatformServices.Tests.Services
     using System.IO;
     using System.Reflection;
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices;
-     using MSTestAdapter.TestUtilities;
+    using MSTestAdapter.TestUtilities;
 #pragma warning disable SA1649 // SA1649FileNameMustMatchTypeName
 
     [TestClass]
