+++ test/UnitTests/PlatformServices.Shared.Unit.Tests/netstandard1.3/Utilities/ns13ReflectionUtilityTests.cs
@@ -1,23 +1,28 @@
 ﻿// Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
-namespace MSTestAdapter.PlatformServices.Desktop.UnitTests.Utilities
+namespace MSTestAdapter.PlatformServices.Tests.Utilities
 {
+#if NETCOREAPP1_0
+    using Microsoft.VisualStudio.TestTools.UnitTesting;
+#else
     extern alias FrameworkV1;
 
+    using Assert = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
+    using CollectionAssert = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.CollectionAssert;
+    using TestClass = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute;
+    using TestMethod = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
+#endif
     using System;
     using System.Collections.Generic;
     using System.Reflection;
 
     using Microsoft.VisualStudio.TestPlatform.MSTestAdapter.PlatformServices.Utilities;
 
-    using Assert = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.Assert;
-    using CollectionAssert = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.CollectionAssert;
-    using TestClass = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestClassAttribute;
-    using TestMethod = FrameworkV1::Microsoft.VisualStudio.TestTools.UnitTesting.TestMethodAttribute;
-
     [TestClass]
+#pragma warning disable SA1649 // File name must match first type name
     public class ReflectionUtilityTests
+#pragma warning restore SA1649 // File name must match first type name
     {
         private ReflectionUtility reflectionUtility;
 
@@ -192,20 +197,6 @@ namespace MSTestAdapter.PlatformServices.Desktop.UnitTests.Utilities
             Assert.AreEqual(2, attribs.Length);
 
             var expectedAttribs = new string[] { "DummyA : a", "DummyA : ba" };
-            CollectionAssert.AreEqual(expectedAttribs, GetAttributeValuePairs(attribs));
-        }
-
-        [TestMethod]
-        public void GetSpecificCustomAttributesOnAssemblyShouldReturnAllAttributes()
-        {
-            var asm = typeof(DummyTestClass).Assembly;
-
-            var attribs = this.reflectionUtility.GetCustomAttributes(asm, typeof(DummyAAttribute));
-
-            Assert.IsNotNull(attribs);
-            Assert.AreEqual(2, attribs.Length);
-
-            var expectedAttribs = new string[] { "DummyA : a1", "DummyA : a2" };
             CollectionAssert.AreEqual(expectedAttribs, GetAttributeValuePairs(attribs));
         }
